--- a/ecc/bls12-377/fr/gkr/gkr_test.go
+++ b/ecc/bls12-377/fr/gkr/gkr_test.go
@@ -113,7 +113,7 @@ func TestSumcheckFromSingleInputTwoIdentityGatesGateTwoInstances(t *testing.T) {
 	pool := polynomial.NewPool(256, 1<<11)
 	workers := utils.NewWorkerPool()
 	o.pool = &pool
-	o.workers = &workers
+	o.workers = workers
 
 	claimsManagerGen := func() *claimsManager {
 		manager := newClaimsManager(circuit, assignment, o)