On my M1 laptop for bn254 gkr+mimc benchmarks (didn't test on the hpc machine, can you do it @Tabaie  ? )

```
benchmark                 old ns/op       new ns/op       delta
BenchmarkGkrMimc19-10     14436017333     10378219291     -28.11%
BenchmarkGkrMimc17-10     3835612209      2801590584      -26.96%

benchmark                 old allocs     new allocs     delta
BenchmarkGkrMimc19-10     2822836        4238771        +50.16%
BenchmarkGkrMimc17-10     1852294        2145689        +15.84%

benchmark                 old bytes      new bytes      delta
BenchmarkGkrMimc19-10     4448614064     2635133856     -40.77%
BenchmarkGkrMimc17-10     1152920304     696349192      -39.60%
```